platform deleter
I added a limiter to a number of platforms that can exist in a scene at the same time
So basically now it deletes all the platforms below the camera! :+1:
Also I removed spiked platform from the platform generator since I thought it's a bad Idea to leave it in the game! :gb: